added favorite spells, fixed cast time error

+++ src/app/services/info-pane.service.ts
@@ -4,13 +4,15 @@ import ICondition from '../models/condition.model';
 import IMeritRef from '../models/meritRef.model';
 import ISpell from '../models/spell.model';
 import ITilt from '../models/tilt.model';
+import ICastConfig from '../models/castConfig.model';
 
 @Injectable({
   providedIn: 'root'
 })
 export class InfoPaneService {
 
-  constructor() { }
+  constructor() {
+  }
 
   currentTab = "";
   visible = false;
@@ -26,12 +28,72 @@ export class InfoPaneService {
     "description": "",
     "arcanum": ""
   };
+  defaultCastingConfig: ICastConfig = {
+    spell: this.spell,
+    // Dice Info
+    freeReach: 0,
+    reach: 0,
+    dicePool: 0,
+    mana: 0,
+    paradoxDesc: 'No Paradox Roll',
+    paradoxDice: 0,
+    // Caster Info
+    showCasterInfo: true,
+    gnosis: 0,
+    highestArcanum: '',
+    charArcanaDots: 0,
+    activeSpells: 0,
+    rulingArcana: true,
+    // Spell Info
+    showSpellInfo: true,
+    requiredArcanum: 0,
+    primaryFactor: 'Potency',
+    isRote: false,
+    isPraxis: false,
+    roteSkillDots: 0,
+    spendWillpower: false,
+    additionalDice: 0,
+    grimoire: false,
+    // Factors
+    potencyValue: 1,
+    advancedPotency: false,
+    range: 0,
+    rangeAdvanced: false,
+    sympatheticRange: false,
+    temporalRange: false,
+    castTime: 0,
+    castingAdvanced: false,
+    timeInBottle: false,
+    scale: 0,
+    scaleAdvanced: false,
+    duration: 0,
+    durationAdvanced: false,
+    extraReach: 0,
+    // Yantras
+    activeYantras: [],
+    yantraBonus: 0,
+    showActiveYantras: true,
+    // Paradox
+    inured: false,
+    previousParadox: 0,
+    sleeperWitness: false,
+    dedicatedToolIndex: 7,
+    additionalMana: 0,
+    numberOfSleepers: 'Straight Roll',
+    // Summary
+    summaryCastTime: '',
+    summaryDuration: '',
+    summaryRange: '',
+    summaryScale: ''
+  };
+  castingConfig?: ICastConfig = undefined;
   attainment: IAttainment = { arcanum: '', dots: 0, name: '', desc: '', system: '' };
   merit: IMeritRef = { name: '', type: '', variableCost: false, minCost: null, maxCost: 0, prerequisite: '', effect: '' };
   tilt: ITilt = { name: '', preDesc: '', effect: ' ', cause: ' ', end: ' ' };
   condition: ICondition = { name: '', effect: '', exampleSkills: '', possibleSources: '', resolution: '', beat: '' };
   meritDots = 0;
   index = 0;
+
 
 
   openPane() {
@@ -59,10 +121,13 @@ export class InfoPaneService {
     this.merit = { name: '', type: '', variableCost: false, minCost: null, maxCost: 0, prerequisite: '', effect: '' };
     this.meritDots = 0;
     this.index = 0;
+    this.castingConfig = undefined;
   }
 
-  setInfoPane(id: string) {
-    this.destroyPanes();
+  setInfoPane(id: string, keepPaneAlive?:boolean) {
+    if(keepPaneAlive== undefined){
+      this.destroyPanes();
+    }
     this.currentTab = id;
     if (!this.visible) {
       this.visible = true;
